Auto Mapper Update

+++ aspnet-core/src/TGC.CMS.Application/CMS_Post/IPostAppService.cs
@@ -3,7 +3,7 @@ using TGC.CMS.CMS_Post.Dtos;
 
 namespace TGC.CMS.CMS_Post
 {
-    public interface IPostAppService : IAsyncCrudAppService<PostDto, long, PagedPostResultRequestDto, CreatePostDto, PostDto>
+    public interface IPostAppService : IAsyncCrudAppService<PostDto, int, PagedPostResultRequestDto, CreatePostDto, PostDto>
     {
 
     }
